Added /search/?query= routing
App.js
- Added /search route
- Added optional /search/?query= route

SearchResults.js
- Created dynamic search results page
- Routes browser to /search/?query=xyz on submit
- Runs search query on page load, nav update

homepage.js
- Added temporary navigate to searchpage
